Adapters and infrastructure moved to application package
