--- next/adapter/presenterlistsellers/presenter_test.go
@@ -1,1 +1,0 @@
-package presenterlistsellers_test
