+++ next/application/adapter/presenterlistsellers/presenter_test.go
@@ -1,0 +1,1 @@
+package presenterlistsellers
